Update Store.py
Learning how to refactor code for re-usability. Trying to create functions for things that can be re-used over and over again. One that I am still having difficulty wrapping my head around is the money tracking/wallet functionality. At the moment, it is a local variable inside the store function just to get the code to a working spot so I can test it, but I'd like to create a Class perhaps that has wallet-type functions that perform those actions in my next update.